+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 ***
 
+# Version 1.0.1 - February 2024
+
+- Update linked data for Guidance extension
+- Log errors for Variant ingest
+- Sort Manifest to have metadata first
+
+***
+
 # Version 1.0.0 - February 2024
 
 - Initial release